backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djoser==2.1.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
Pillow==9.0.0
django-filter==24.2
python-dotenv==1.0.1
gunicorn==20.1.0
